requirements.txt
django==4.2.1
CloudFlare~=2.11.1
mysql-connector-python~=8.0.33
openai~=0.27.6
requests~=2.28.2
colorama~=0.4.6
python-slugify~=8.0.1